The Hough box segmentation seems like a dead end
Realized when applying the neural net that the round-off errors
and non-linear errors becomes too significant when segmenting
boxes. The segmentation seems to work fine for boxes close to
origin, but gets worse as one moves away from the origin.

Going to leave this approach for now but keep the sanity checks
and try to segment the digits alone in a similar manner as I did
for the board extraction (using thresholded contours).

Signed-off-by: Alf-Rune Siqveland <alf.rune@siqveland.no>